+++ src/selectorValueChecker.ts
@@ -1,52 +1,55 @@
 import { Logging } from 'homebridge';
-import puppeteer from 'puppeteer-core';
+import { chromium } from 'playwright-core';
 import { runStepsBeforeCheck } from './runStepsBeforeCheck';
 import { ChangeCheck } from './types/optionTypes';
 
 require('dotenv').config();
 
 type getValueFromPageProps = {
-    executablePath: string;
+    browserPath: string;
     changeCheck: ChangeCheck;
     waitForSelectorTimeout?: number;
     log: Logging;
     verboseLogging?: boolean;
-    previousValue?: string;
 };
 
 export const getValueFromPage = async (props: getValueFromPageProps) => {
     let foundValue: string | undefined = undefined;
-    const browser = await puppeteer.launch({
-        executablePath: props.executablePath,
-        args: ['--no-sandbox', '--disable-setuid-sandbox']
+    const browser = await chromium.launch({
+        executablePath: props.browserPath
     });
 
     try {
         if (props.verboseLogging) { props.log('Initialize browser'); }
 
         const page = await browser.newPage();
+
+        if (process.env.NODE_ENV === 'test') {
+            page.setDefaultTimeout(2000);
+        }
+        
         if (props.verboseLogging) { props.log('Browser initialized'); }
-        await page.goto(props.changeCheck.url, { waitUntil: 'networkidle2' });
+        
+        await page.goto(props.changeCheck.url, { waitUntil: 'networkidle' });
 
         // Execute steps before actual check
         if (props.changeCheck.stepsBeforeCheck) {
             await runStepsBeforeCheck(page, props.changeCheck.stepsBeforeCheck);
         }
 
-        await page.waitForSelector(props.changeCheck.selector, {
+        const text = await page.innerText(props.changeCheck.selector, {
             timeout: props.waitForSelectorTimeout || 30000 // 30 seconds
         });
 
         if (props.verboseLogging) { props.log('Selector loaded'); }
 
-        const element = await page.$(props.changeCheck.selector);
-        const text = await page.evaluate(element => { return element.textContent; }, element);
-
-        foundValue = text.replace(/(?:\r\n|\r|\n)/g, '').replace(/\s+/g, ' ').trim();
+        if (text) {
+            foundValue = text.replace(/(?:\r\n|\r|\n)/g, '').replace(/\s+/g, ' ').trim();
+        }
     } catch (e) {
         props.log.warn(e.toString());
 
-        foundValue = props.previousValue;
+        foundValue = undefined;
     } finally {
         await browser.close();