+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2021-03-13
+### Added
+- Added `click` as a `stepsBeforeCheck` action
+- Added travis.yml for CI
+### Changed
+- Migrated from `puppeteer` to `playwright` due to less code writing and seemingly better stability. 
+- Abstracted some plugin code from Homebridge to make it better testable.   
+- Improved the config.schema.json to make a more intuitive settings UI.
+
 ## [1.0.5] - 2021-03-13
 ### Added
 - `stepsBeforeCheck` option for every changeCheck. See README.md for how to use it.